# elm-frontend/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "吃饱了吗点餐平台",
    "description": "Tianjin University Software Engineering Project",
    "version": "1.0.0"
  },
...
  "paths": {
    "/api/businesses/{id}": {
      "get": {
        "summary": "getBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "getBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "updateBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "updateBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "deleteBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "patch": {
        "summary": "patchBusiness",
        "deprecated": false,
        "description": "",
        "operationId": "patchBusiness",
        "tags": ["管理店铺"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Business"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultBusiness"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/foods/{id}": {
      "get": {
        "summary": "返回查询到的一条商品记录",
        "deprecated": false,
        "description": "",
        "operationId": "getFoodById",
        "tags": ["管理商品"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultFood"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "更新商品信息",
        "deprecated": false,
        "description": "更新指定ID的商品信息",
        "operationId": "updateFood",
        "tags": ["管理商品"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "商品ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Food"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultFood"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除商品",
        "deprecated": false,
        "description": "删除指定ID的商品",
        "operationId": "deleteFood",
        "tags": ["管理商品"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "商品ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/carts": {
      "get": {
        "summary": "获取当前用户的购物车",
        "deprecated": false,
        "description": "获取当前登录用户的所有购物车项目",
        "operationId": "getCurrentUserCart",
        "tags": ["管理购物车"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultListCart"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "addCartItem",
        "deprecated": false,
        "description": "",
        "operationId": "addCartItem",
        "tags": ["管理购物车"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultCart"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/carts/{id}": {
      "patch": {
        "summary": "更新购物车项目",
        "deprecated": false,
        "description": "更新购物车中某个项目的数量",
        "operationId": "updateCartItem",
        "tags": ["管理购物车"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "购物车项目ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "format": "int32"
                  }
                },
                "required": ["quantity"]
              },
              "examples": {}
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/HttpResultCart"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除购物车项目",
        "deprecated": false,
        "description": "从购物车中删除一个项目",
        "operationId": "deleteCartItem",
        "tags": ["管理购物车"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "购物车项目ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    },
                    "data": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "food": {
            "$ref": "#/components/schemas/Food"
          },
          "customer": {
            "$ref": "#/components/schemas/User"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Authority": {
        "required": ["name"],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "Business": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "businessName": {
            "type": "string"
          },
          "businessOwner": {
            "$ref": "#/components/schemas/User"
          },
          "businessAddress": {
            "type": "string"
          },
          "businessExplain": {
            "type": "string"
          },
          "businessImg": {
            "type": "string"
          },
          "orderTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "startPrice": {
            "type": "number"
          },
          "deliveryPrice": {
            "type": "number"
          },
          "remarks": {
            "type": "string"
          }
        },
        "required": ["id"]
      },
...
      "User": {
        "required": ["username"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "username": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "authorities": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Authority"
            }
          }
        }
      },
      "HttpResultBusiness": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Business"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Food": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time"
          },
          "creator": {
            "type": "integer",
            "format": "int64"
          },
          "updater": {
            "type": "integer",
            "format": "int64"
          },
          "deleted": {
            "type": "boolean"
          },
          "foodName": {
            "type": "string"
          },
          "foodExplain": {
            "type": "string"
          },
          "foodImg": {
            "type": "string"
          },
          "foodPrice": {
            "type": "number"
          },
          "business": {
            "$ref": "#/components/schemas/Business"
          },
          "remarks": {
            "type": "string"
          }
        }
      },
      "HttpResultFood": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Food"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "HttpResultCart": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Cart"
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "HttpResultListCart": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cart"
            }
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}